pyproject.toml
[tool.poetry]
name = "parlati"
version = "0.1.0"
description = "Q&A Bot with general information on how to live in Ticino"
authors = ["Luca Simone <info@lucasimone.info>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
scrapy = "^2.9.0"
beautifulsoup4 = "^4.12.2"
openai = "^0.27.8"
langchain = "^0.0.200"
bleach = "^6.0.0"
html2text = "^2020.1.16"
python-slugify = "^8.0.1"
chromadb = "^0.3.26"
tiktoken = "^0.4.0"
python-telegram-bot = "^20.3"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
